Cleanup util/math BUILD file and ensure clean compilation and run.
- Fix any unused variables and missing dependencies.
- Add ASAN config to .bazelrc and ensure clean run.
- Fix -Wcompare-unsigned warning in exactfloat.cc

Now builds and runs cleanly even with ASAN on (only ASAN notification is
internal to Abseil).

+++ src/s2/util/math/exactfloat/exactfloat.cc
@@ -371,7 +371,7 @@ int ExactFloat::GetDecimalDigits(int max_digits, std::string* digits) const {
   std::string::size_type pos = digits->find_last_not_of('0') + 1;
   bn_exp10 += digits->size() - pos;
   digits->erase(pos);
-  ABSL_DCHECK_LE(digits->size(), max_digits);
+  ABSL_DCHECK_LE(static_cast<int>(digits->size()), max_digits);
 
   // Finally, we adjust the base-10 exponent so that the mantissa is a
   // fraction in the range [0.1, 1) rather than an integer.
